Use Redux to track weather we are loading notes
Previously the rest client was emitting events when it started to load
and when it received responses for network calls polling the notes.

`<NotesList />` subscribed to these events in order to show a loading
spinner in the UI. Those subscriptions have been removed and now it uses
Redux as the communications mechanism.