Use older Redux store enhancer API
Although the newer API version is cleaner, Calypso still uses redux@<3.1
where the new API was introduced. The newer versions of Redux still
allow for this old syntax so this change uses the common denominator
where it will work in both places.

Previously the middleware simply didn't load in Calypso