Properly change isShowing (#45)
Previously we were triggering the state changes from showing to hidden
from inside the app but this can only be controlled by the host app. We
can _request_ those changes but cannot actually force them.

Now, we only update the internal state when the top-level React
component detects a change. All other actions to open or close are
requests to the outside world.