Preserve list view reset on iframe view
Previously when working inside an iframe, each time the app opened it
would reset to the list view. With the updates we have made it tries to
preserve the previous state: detail view/list view/scroll
position/etc...

This is great for directly embedding as a React component but it was
causing an issue on existing iframe updates. If the app were in detail
view/wide view when closed it would try to return when opened back to
the detail view but lose the widescreen state. This was because of a
state reset on the wide view.

Since we can't go back and update all existing Jetpack deployments we're
trying to preserve this weird behavior. With this patch we will actually
be slightly changing (the scroll positions will be roughly the same as
when closed) but keeping the reset-to-list-view consistent. Calypso will
abandon the existing undesirable behavior but iframes will preserve
them.