Stop clearing the selected note when closing panel
Previously whenever we closed the notes panel we would also unselected
the selected note. This meant that we introduced some funny behavior
when closing with the escape key vs with the `n` key or by clicking on
the bell. When closing from the outside the `closePanel()` doesn't get
called and thus the note is not reset. On the other hand, when closing
with the escape key that actions _does_ run.

Now the two behaviors should be unified.