Fix logic bug in settings placement and refactor time groups
With the original commit in this branch the settings icon was being
placed on the 'Today' group if it existed, but if no notes were
available in that group then the icon wasn't added.

This commit changes the behavior so that the settings icon appears on
the first visible time group header.

Further, some of the time grouping was refactored for simplicity and to
remove mutation. Most notably the `timeBoundaries` have been adjusted so
that they form real boundaries (with proper infinite ends) and zipped
into pairs of boundaries of time values for `(before, after)` referring
to the range of times within them.