Simplify ListHeader and move window open to middleware
The `<ListHeader />` didn't need any local state so it has been
converted into a functional stateless component.

The action to open the settings view has also been turned into a plain
action and the middleware is picking it up to open a new window.

Future work should intercept that in Calypso to navigate Calypso to the
settings pane instead of directly opening a new window.