Remove unused prop-passers for "message passing" (#123)
Previously we have had a hard time managing communication between the
parent app/iframe and the notifications panel. Props such as
`onToggleLayout` and `appRefresher` were introduced to help this but
they were the wrong solution and lead to even more confusion.

The custom Redux store enhancer injection is the right solution and this
change removes the uses of the old prop passers so that we can start to
bring more declarative and centralized points of interaction between the
inside of the app and the outside world.