Fix: Set layout as side-effect
We have previously been setting the layout when clicking
on a notification from the summary list. This set the
widescreen status which would open and close the detail
drawer.

This wasn't triggering however when opening a notification
via the keyboard shortcuts. At least, the drawer wouldn't
open on account of not notifying to the parent window
which screen width setting was set.

This entire system probably needs to be replaced with
actions like `OPEN_DRAWER` or `OPEN_DETAIL_VIEW` or
something like that. We have significant overlap with
`selectedNoteId`

This patch uses a middleware watcher to set the layout
whenever a note is selected or unselected so that it
won't matter how the change happened, just that it did.