Fix: Note reads across browser sessions, <Note> cleanup
Resolves #130

In the earlier versions of the app we used stateful updates to the React
render and relied on storage events to communicate between sessions that
notes were being read. This led to re-renders reflecting those changes
by listening to the `ready` event from `rest-client`.

With the introduction of Redux and the removal of those event-emitters
we lost this functionality and notes weren't being marked read. Further,
because we were still tied to the _impure_ render modes the note
component wasn't updating even when state was.

Now, upon retrieval of such a storage message we are marking the note as
read via a dispatched action. The relevant components are now no longer
marked as _impure_ and will update on all state changes.

Interestingly enough an old `shouldComponentUpdate` functionw was
preventing updates because it wasn't comparing against `getNoteIsRead()`
(since it's now separate from the note object).