Styling: Center all images by default
A couple times recently we have seen attempts to center images through
styles embedded in notification content directly or by using a "badge"
type notification. The problem with these approaches is that they end
up working around the system in the notifications panel and become
fragile and coupled to things that are implementation details instead
of design elements.

After discussion with a few people we realized that there is no reason
for us not to automatically center images displayed in the
notifications themselves. In fact, the mobile clients already do this.

In this patch we're centering all images inside of the notifications.
This should make it possible to do what people have wanted, which is
to have a left-aligned paragraph with a centered image which is also
not a badge/trophy notification.